requirements.txt
numpy==1.26.1
tensorflow==2.14.0
pyaudio==0.2.13
matplotlib==3.7.2
seaborn
librosa
pydub
SpeechRecognition
gtts
openai
paho-mqtt
openai-whisper